services/authentication/requirements.txt
fastapi
python-multipart
pymongo
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
redis